+++ winpr/libwinpr/thread/thread.c
@@ -576,7 +576,7 @@ static BOOL winpr_StartThread(WINPR_THREAD* thread)
 	pthread_attr_setdetachstate(&attr, PTHREAD_CREATE_JOINABLE);
 
 	if (thread->dwStackSize > 0)
-		pthread_attr_setstacksize(&attr, (size_t)thread->dwStackSize);
+		pthread_attr_setstacksize(&attr, thread->dwStackSize);
 
 	thread->started = TRUE;
 	reset_event(thread);
